Move test utils to util
